add cards to homepage
