+++ src/components/stylesheets/Song.module.css
@@ -1,0 +1,12 @@
+.card-img {
+  max-width: 180px;
+}
+.card-text {
+  max-width: 150px;
+  display: flex;
+  justify-content: flex-end;
+}
+
+.card {
+  max-height: 280px;
+}
